backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langgraph
supabase
pydantic
python-dotenv
requests
pillow
moviepy
tiktoken
numpy
pandas
python-multipart
jinja2
aiofiles
httpx
beautifulsoup4
selenium
undetected-chromedriver
pytube
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
apscheduler
celery
redis
groq
